+++ okkh/src/main/java/com/kh/okkh/chat/model/vo/ChatRoom.java
@@ -10,7 +10,6 @@ import java.util.UUID;
 import org.springframework.web.socket.WebSocketSession;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -22,23 +21,23 @@ import lombok.NoArgsConstructor;
 @Setter
 @ToString
 public class ChatRoom {
-		private int roomNo;
-		private String roomId;
-		private String roomTitle;
-		private Date createDate;
-		private String status;
-		private String lastChat;
-		private Date lastDate;
-		private int fileNo;
-	
+	private int roomNo;
+	private String roomTitle;
+	private Date createDate;
+	private String status;
+	private String lastChat;
+	private Date lastDate;
+	private int fileNo;
+
+	private String roomId;
 	private String name;
 	private Set<WebSocketSession> sessions = new HashSet<>();
 	
 	public static ChatRoom create(String name) {
-		ChatRoom chatRoom = new ChatRoom();
-		chatRoom.roomId = UUID.randomUUID().toString();
-		chatRoom.name = name;
-		return chatRoom;
+		ChatRoom room = new ChatRoom();
+		room.roomId = UUID.randomUUID().toString();
+		room.name = name;
+		return room;
 	}
 		
 }
