+++ okkh/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -32,7 +32,6 @@
 	
 	<context:component-scan base-package="com.kh.okkh" />
 	
-	
 	<beans:bean id="chatWebsocketHandler" class="com.kh.okkh.chat.controller.ChatWebsocketHandler"/>
 	
 	<websocket:handlers>
